// src/typings/api/assets/assets.d.ts
 export type CoinListResp =  {
  /** 币种 app 端 logo */
  appLogo?: any;
  /** 币种 web 端 logo */
  webLogo?: any;
  /** 币种名称 */
  coinName?: string;
  /** 币全名 */
  coinFullName?: string;
  /** 可用数量 */
  availableAmount?: string;
  /** 冻结数量 */
  lockAmount?: string;
  /** 总数量 */
  totalAmount?: string;
  /** 币种 id */
  coinId: string;
  /** 下单锁定  */
  orderLockAmount?: string;
  /** BTC 估值 */
  btcAmount?: number;
  /** usdt 估值 */
  usdtAmount?: number;
  /** usd 估值 */
  usdBalance?: number;
  /** 币种 */
  symbol: string;
  /** 仓位数量  */
  positionAmount?: string
}
// ...
 export type AssetsTradeListResp = {
  spot?: TradeListSpotResp[]
  /** 合约币对列表 */
  perpetual?: TradeListPerpetualResp[]

 }

 export type TradeListSpotResp = {
  id?: number;
  /** 标的币 */
  sellCoinId?: number;
  /** 计价币 */
  buyCoinId?: number;
  /** 交易区 */
  tradeArea?: number;
  symbolName?: string;
  /** 涨跌幅 */
  chg?: string;
  /** 最高价 */
  high?: string;
  /** 当前价 */
  last?: string;
  /** 最低价 */
  low?: string;
  /** 开盘价 */
  open?: string;
  /** 成交量 */
  volume?: string;
  /** 成交额 */
  quoteVolume?: string;
  time?: number;
  /** 标的币 名 */
  baseSymbolName?: string;
  /** 计价币 名 */
  quoteSymbolName?: string;
  /** 钱包币对名 */
  symbolWassName?: string;
  /** 排序字段 */
  sort?: number;
 }

 export type TradeListPerpetualResp = {
  /** 标的币名 */
  /** 标的币精度 */
  baseSymbolScale?: number;
  /** 计价币精度 */
  quoteSymbolScale?: number;
  /** 最小下单数量 */
  minCount?: number;
  /** 是否添加自选 1=自选 */
  favourite?: number;
  /** 币对名 */
  symbolName?: string;
  /** 最低价 */
  low?: string;
  /** 开盘价 */
  open?: string;
  plusFeeRate?: string;
  /** 最高价 */
  high?: string;
  sellFee?: string
  /** 是否开盘;1=开盘中，2=停盘中，3=预约开盘 */
  marketStatus?: number;
  tradeArea?: string
  /** 是否删除 2=未删除 1=已删除 */
  isDelete?: number;
  /** wass 钱包名 */
  symbolWassName?: string;
  /** 下单数量偏移 */
  amountOffset?: string;
  /** 标的币 id */
  sellCoinId?: number;
  /** 下单价格偏移 */
  priceOffset?: number;
  /** 商户 id */
  businessId?: number;
  /** 流通量 */
  circulatingSupply?: number;
  /** 计价币名 */
  quoteSymbolName?: string;
  /** 计价币 id */
  buyCoinId?: number;
  id?: number;
  /** 结算规则 */
  settlementRules?: number;
  /** 涨跌幅 */
  chg?: string;
  /** 标的币名 */
  baseSymbolName?: string;
  /** 成交量 */
  volume?: string;
  /** 合约类型 delivery =交割 ,perpetual =永续 */
  typeIndex?: string;
  /** 最新价 */
  last?: string;
  /** 最大下单金额 */
  maxAmount?: string;
  /** 最大下单数量 */
  maxCount?: string;
  /** 成交额 */
  quoteVolume?: string;
  /** 排序字段 */
  sort?: number;
  /** 时间 */
  time?: number;
  /** 结算时间 */
  cycle?: string;
  /** 1=启用 2=禁用 */
  enableState?: number;
  /** 最小下单金额 */
  minAmount?: string;
  /** 1=交易 2=不交易 */
  isShare?: number;
  minusFeeRate?: string;
  
 }